feat(folders): when downloading stories, they get stored in folders; also, working TOC
